+++ src/rust/src/rdataframe/read_parquet.rs
@@ -39,7 +39,8 @@ pub fn new_from_parquet(
         rechunk,
         row_count,
         low_memory,
-        cloud_options: None, //TODO implement cloud options
+        cloud_options: None,  //TODO implement cloud options
+        use_statistics: true, //TODO expose use statistics
     };
 
     let lf_result = pl::LazyFrame::scan_parquet(path, args)